requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.51.0
accelerate>=0.20.0
safetensors>=0.3.0
Pillow>=9.0.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
opencv-python>=4.5.0
tqdm>=4.62.0
requests>=2.25.0
huggingface-hub>=0.16.0
tokenizers>=0.13.0
sentencepiece>=0.1.96
protobuf>=3.20.0
